requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.200

# MiDaS depth estimation
timm>=0.9.0

# Data / numerics
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Logging
wandb>=0.16.0
tensorboard>=2.14.0

# Visualisation / plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Utils
PyYAML>=6.0
tqdm>=4.66.0
pandas>=2.0.0

# Ensemble / WBF
ensemble-boxes>=1.0.9
